requirements.txt
streamlit==1.32.2
langchain==0.0.353
langchain-groq==0.0.1
groq==0.4.1
sqlalchemy==2.0.25
python-dotenv==1.0.1
faiss-cpu==1.7.4
httpx==0.27.0
